backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-chess==1.999
openai==1.35.3
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0
